requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.0.6
django-crispy-forms==1.9.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==7.1.2
psycopg2==2.8.4
pycodestyle==2.5.0
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-decouple==3.3
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
var-dump==1.2
whitenoise==5.0.1
wrapt==1.12.1
